Added processing of OBJECT and FLUX, improved bad pixel removal and more
Added a function process_object_frames and a function process_flux_frames. Both function still need proper centering routines implemented.

Added that sigma-filtering no happens iterative until there is no more bad pixel to be removed. Also adjusted the kernel which has improved the bad pixel removing.

Added the possibility to save the pre-processed single-sum and -difference image cubes, and reading them again to skip the pre-processing.

Removed the variable recompute as we can now save the single sum and single difference images to be used in post processing. Centering is the most difficult part, so the one thing you want to re-run is the centering so no need to save the center coordinates then.