Added centering methods + much more
Added centering methods using center frames, 2D Gaussian fitting, cross-correlation (registering) and manual. The program outputs images and files to check results. With center frames you have the option to subtract the OBJECT frame closest in time to reduce noise. With Gaussian fitting you can mask saturated pixels and replace the background with Gaussian noise. Registering fits the first image with a Gaussian and then does the cross-correlation to align the rest of the images.