Added a parameter param_centering_satellite_spots
Added a parameter param_centering_satellite_spots to distinguish between the fitting of the satellite spots of CENTER-frames and the fitting of the PSF position in OBJECT-frames with a Gaussian or with cross-correlation. When using a large enough crop_radius for param_centering_object (used for OBJECT-frames) the fitting is robust for non-coronagraphic data, as long as there is no brighter source in the image.